Cast with care, please!

+++ src/cogs/render.py
@@ -108,9 +108,9 @@ class Renderer:
 		'''Apply rgb color multiplication (0-255)'''
 		r,g,b = rgb
 		rc,gc,bc,ac = sprite.split()
-		rc = rc.point(lambda i: i * (r/256))
-		gc = gc.point(lambda i: i * (g/256))
-		bc = bc.point(lambda i: i * (b/256))
+		rc = rc.point(lambda i: int(i * (r/256)))
+		gc = gc.point(lambda i: int(i * (g/256)))
+		bc = bc.point(lambda i: int(i * (b/256)))
 		
 		return Image.merge('RGBA', (rc,gc,bc,ac))
